requirements.txt
streamlit==1.45.0
streamlit-webrtc==0.62.4
openai==1.84.0
openai-whisper==20240930
torch==2.7.1
pydub==0.25.1
moviepy==2.2.1
python-dotenv==1.1.0
tiktoken==0.9.0
docx2txt==0.9